reimplementado casi toda la conexion, ahora hay causalconn y causalserver, un servidor mas rapido que mosquitto, ademas añadido otras metricas

+++ tests/test_consistency.sh
@@ -1,18 +1,25 @@
 #!/bin/sh
 echo "CONSISTENCY TEST - 3 REPLICAS"
 echo "Compiling test..."
-go build ./test_consistency.go
+if ! go build ./test_consistency.go; then
+	echo "Compilation error"
+	exit 1
+fi
 
 rm -f log0.txt log1.txt log2.txt
-# MQTT=mqtt://140.238.237.68:1883
-MQTT=mqtt://broker.hivemq.com:1883
+
+SERVER=localhost:5000
+# SERVER=140.238.237.68:1883
+# SERVER=mqtt://test.mosquitto.org:1883
+# SERVER=mqtt://mqtt.eclipseprojects.io:1883
+# SERVER=mqtt://broker.hivemq.com:1883
 
 echo "Run Replica 0"
-./test_consistency.exe 0 "$MQTT" > log0.txt &
+./test_consistency.exe 0 "$SERVER" > log0.txt &
 echo "Run Replica 1"
-./test_consistency.exe 1 "$MQTT" > log1.txt &
+./test_consistency.exe 1 "$SERVER" > log1.txt &
 echo "Run Replica 2"
-./test_consistency.exe 2 "$MQTT" > log2.txt &
+./test_consistency.exe 2 "$SERVER" > log2.txt &
 
 echo "Waiting for CRDTs eventual consistency..."
 wait $(jobs -p)
